Move the buffer JFR events into the buffer package and make them package-private
