+++ buffer/src/main/java/io/netty/buffer/AbstractChunkEvent.java
@@ -13,10 +13,8 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.buffer.jfr;
+package io.netty.buffer;
 
-import io.netty.buffer.AbstractAllocatorEvent;
-import io.netty.buffer.AbstractByteBuf;
 import io.netty.util.internal.UnstableApi;
 import jdk.jfr.DataAmount;
 import jdk.jfr.Description;
@@ -24,30 +22,22 @@ import jdk.jfr.MemoryAddress;
 
 @UnstableApi
 @SuppressWarnings("Since15")
-public abstract class AbstractBufferEvent extends AbstractAllocatorEvent {
+abstract class AbstractChunkEvent extends AbstractAllocatorEvent {
     @Description("The type of allocator this event is for")
     public AllocatorType allocatorType;
     @DataAmount
-    @Description("Configured buffer capacity")
-    public int size;
-    @DataAmount
-    @Description("Actual allocated buffer capacity")
-    public int maxFastCapacity;
-    @DataAmount
-    @Description("Maximum buffer capacity")
-    public int maxCapacity;
-    @Description("Is this buffer referencing off-heap memory?")
+    @Description("Size of the chunk")
+    public int capacity;
+    @Description("Is this chunk referencing off-heap memory?")
     public boolean direct;
     @Description("The memory address of the off-heap memory, if available")
     @MemoryAddress
     public long address;
 
-    public void fill(AbstractByteBuf buf, AllocatorType allocatorType) {
+    public void fill(ChunkInfo chunk, AllocatorType allocatorType) {
         this.allocatorType = allocatorType;
-        size = buf.capacity();
-        maxFastCapacity = buf.maxFastWritableBytes() + buf.writerIndex();
-        maxCapacity = buf.maxCapacity();
-        direct = buf.isDirect();
-        address = getMemoryAddressOf(buf);
+        capacity = chunk.capacity();
+        direct = chunk.isDirect();
+        address = chunk.memoryAddress();
     }
 }
