+++ buffer/src/main/java/io/netty/buffer/FreeChunkEvent.java
@@ -13,20 +13,22 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.buffer.jfr;
+package io.netty.buffer;
 
 import io.netty.util.internal.UnstableApi;
-import jdk.jfr.DataAmount;
+import jdk.jfr.Category;
 import jdk.jfr.Description;
 import jdk.jfr.Enabled;
+import jdk.jfr.Label;
 
 @UnstableApi
 @Enabled(false)
 @SuppressWarnings("Since15")
-@Description("Triggered when a buffer is reallocated for resizing in an allocator. " +
-        "Will be followed by an AllocateBufferEvent")
-public final class ReallocateBufferEvent extends AbstractBufferEvent {
-    private static final ReallocateBufferEvent INSTANCE = new ReallocateBufferEvent();
+@Category("Netty")
+@Label("Chunk Free")
+@Description("Triggered when a memory chunk is freed from an allocator")
+final class FreeChunkEvent extends AbstractChunkEvent {
+    private static final FreeChunkEvent INSTANCE = new FreeChunkEvent();
 
     /**
      * Statically check if this event is enabled.
@@ -35,7 +37,6 @@ public final class ReallocateBufferEvent extends AbstractBufferEvent {
         return INSTANCE.isEnabled();
     }
 
-    @DataAmount
-    @Description("Targeted buffer capacity")
-    public int newCapacity;
+    @Description("Was this chunk pooled, or was it a one-off allocation for a single buffer?")
+    public boolean pooled;
 }
