--- buffer/src/main/java/io/netty/buffer/jfr/AllocatorType.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2025 The Netty Project
- *
- * The Netty Project licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package io.netty.buffer.jfr;
-
-import io.netty.util.internal.UnstableApi;
-
-/**
- * The type of allocator that an event is related to.
- */
-@UnstableApi
-public enum AllocatorType {
-    /**
-     * The event is produced by the {@link io.netty.buffer.PooledByteBufAllocator}.
-     */
-    pooled,
-    /**
-     * The event is produced by the {@link io.netty.buffer.AdaptiveByteBufAllocator}.
-     */
-    adaptive,
-    /**
-     * The event is produced by an unknown or indeterminate allocator.
-     */
-    unknown
-}
